+++ CHANGELOG.md
@@ -1,9 +1,13 @@
-BQ_UWSGI COOKBOOK CHANGELOG
+tgw_uwsgi COOKBOOK CHANGELOG
 ===========================
+
+1.0.0 (2017-12-22)
+------------------
+- Updated copyright to new owner name.
 
 0.2.0 (2017-12-15)
 ------------------
-- [ISSUE 2](https://github.com/arktos65/bq_uwsgi/issues/2)::Updated Rubocop to 0.52.0
+- [ISSUE 2](https://github.com/arktos65/tgw_uwsgi/issues/2)::Updated Rubocop to 0.52.0
 - Updated Test Kitchen to 1.19.2
 - Updated Kitchen Dokken to 2.6.6